Copy the content of 18-index.html into 20-index.html
in your nav tags
create a link to / with the text Home
create an anchor to services with the text Services
create an anchor to works with the text Works
create an anchor to about with the text About
create an anchor to latest_news with the text Latest news
create an anchor to testimonials with the text Testimonials
create an anchor to contact with the text Contact
For now, the anchor links will not work. We will make them work in the CSS project.